corrección ej1 y realización ej2

+++ ejercicio1.py
@@ -4,5 +4,5 @@ n1= int(input("Dime un número entero: "))
 #Type() para capturar el tipo de dato
 print(n1,(type(n1)))
 #################################
-n2= int(input("Dime un número flotante: "))
+n2= float(input("Dime un número flotante: "))
 print(n2, type(n2))
